Thread Policy

+++ app/Http/Controllers/ThreadsController.php
@@ -114,6 +114,8 @@ class ThreadsController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
+        $this->authorize('update', $thread);
+
         $thread->delete();
 
         if(request()->wantsJson()) {
